+++ coordinator/migrations/2023-12-04-152138_add_position_state_proposed/down.sql
@@ -3,3 +3,5 @@
 -- Postgres does not allow removing enum type values. One can only re-create an enum type with fewer values and replace the references.
 -- However, there is no proper way to replace the values to be removed where they are used (i.e. referenced in `positions` table)
 -- We opt to NOT remove enum values that were added at a later point.
+
+select 1;