+++ coordinator/migrations/2024-01-16-135827_remove_unneeded_fields_from_collab_revert_table/up.sql
@@ -1,0 +1,6 @@
+-- Your SQL goes here
+ALTER TABLE collaborative_reverts DROP COLUMN "funding_txid";
+ALTER TABLE collaborative_reverts DROP COLUMN "funding_vout";
+-- we do not store channels at the moment, hence, we can't have the reference to channels
+ALTER TABLE collaborative_reverts
+    DROP CONSTRAINT IF EXISTS collaborative_reverts_channel_id_fkey;