+++ coordinator/src/admin.rs
@@ -1,6 +1,5 @@
 use crate::collaborative_revert;
 use crate::db;
-use crate::parse_channel_id;
 use crate::parse_dlc_channel_id;
 use crate::routes::AppState;
 use crate::AppError;
@@ -13,8 +12,6 @@ use bdk::FeeRate;
 use bdk::LocalUtxo;
 use bdk::TransactionDetails;
 use bitcoin::secp256k1::PublicKey;
-use bitcoin::OutPoint;
-use commons::CollaborativeRevertCoordinatorExpertRequest;
 use commons::CollaborativeRevertCoordinatorRequest;
 use dlc_manager::channel::Channel;
 use dlc_manager::contract::Contract;
@@ -321,55 +318,16 @@ pub async fn collaborative_revert(
     revert_params: Json<CollaborativeRevertCoordinatorRequest>,
 ) -> Result<(), AppError> {
     let channel_id_hex = revert_params.channel_id.clone();
-    let channel_id = parse_channel_id(channel_id_hex.as_str())
+    let channel_id = parse_dlc_channel_id(channel_id_hex.as_str())
         .map_err(|e| AppError::BadRequest(format!("Invalid channel ID provided: {e:#}")))?;
-
-    let funding_txo = OutPoint {
-        txid: revert_params.txid,
-        vout: revert_params.vout,
-    };
 
     collaborative_revert::propose_collaborative_revert(
         state.node.inner.clone(),
         state.pool.clone(),
         state.auth_users_notifier.clone(),
         channel_id,
-        revert_params.price,
         revert_params.fee_rate_sats_vb,
-        funding_txo,
-    )
-    .await
-    .map_err(|e| {
-        AppError::InternalServerError(format!("Could not collaboratively revert channel: {e:#}"))
-    })?;
-
-    tracing::info!(channel_id = channel_id_hex, "Proposed collaborative revert");
-
-    Ok(())
-}
-
-#[instrument(skip_all, err(Debug))]
-pub async fn expert_collaborative_revert(
-    State(state): State<Arc<AppState>>,
-    revert_params: Json<CollaborativeRevertCoordinatorExpertRequest>,
-) -> Result<(), AppError> {
-    let channel_id_hex = revert_params.channel_id.clone();
-    let channel_id = parse_channel_id(channel_id_hex.as_str())
-        .map_err(|e| AppError::BadRequest(format!("Invalid channel ID provided: {e:#}")))?;
-
-    let funding_txo = OutPoint {
-        txid: revert_params.txid,
-        vout: revert_params.vout,
-    };
-
-    collaborative_revert::propose_collaborative_revert_without_channel_details(
-        state.node.inner.clone(),
-        state.pool.clone(),
-        state.auth_users_notifier.clone(),
-        channel_id,
-        funding_txo,
-        revert_params.coordinator_amount,
-        revert_params.fee_rate_sats_vb,
+        revert_params.counter_payout,
         revert_params.price,
     )
     .await