+++ coordinator/src/position/models.rs
@@ -6,11 +6,10 @@ use anyhow::Result;
 use bitcoin::secp256k1::PublicKey;
 use bitcoin::Address;
 use bitcoin::Amount;
-use bitcoin::Txid;
 use commons::order_matching_fee_taker;
 use commons::TradeParams;
 use dlc_manager::ContractId;
-use lightning::ln::ChannelId;
+use dlc_manager::DlcChannelId;
 use rust_decimal::prelude::Signed;
 use rust_decimal::prelude::ToPrimitive;
 use rust_decimal::Decimal;
@@ -368,15 +367,13 @@ pub fn leverage_short(
 
 #[derive(Clone, Debug)]
 pub struct CollaborativeRevert {
-    pub channel_id: ChannelId,
+    pub channel_id: DlcChannelId,
     pub trader_pubkey: PublicKey,
-    pub price: f32,
+    pub price: Decimal,
     pub coordinator_address: Address,
     pub coordinator_amount_sats: Amount,
     pub trader_amount_sats: Amount,
     pub timestamp: OffsetDateTime,
-    pub txid: Txid,
-    pub vout: u32,
 }
 
 impl std::fmt::Debug for NewPosition {